making the api consistent

+++ libmoat/api/libbarbican.h
@@ -16,16 +16,16 @@
 extern "C" {
 #endif
 
-sgx_status_t SGX_CDECL recv_dh_msg1_ocall(size_t* retval, sgx_measurement_t *target_enclave, sgx_dh_msg1_t* dh_msg1, size_t session_id);
-sgx_status_t SGX_CDECL send_dh_msg2_recv_dh_msg3_ocall(size_t* retval, sgx_dh_msg2_t* dh_msg2, sgx_dh_msg3_t* dh_msg3, size_t session_id);
-sgx_status_t SGX_CDECL send_dh_msg1_recv_dh_msg2_ocall(size_t* retval, sgx_measurement_t *target_enclave, sgx_dh_msg1_t* dh_msg1, sgx_dh_msg2_t* dh_msg2, size_t session_id);
-sgx_status_t SGX_CDECL send_dh_msg3_ocall(size_t* retval, sgx_dh_msg3_t* dh_msg3, size_t session_id);
+sgx_status_t SGX_CDECL recv_dh_msg1_ocall(size_t* retval, sgx_measurement_t *target_enclave, sgx_dh_msg1_t* dh_msg1, int64_t session_id);
+sgx_status_t SGX_CDECL send_dh_msg2_recv_dh_msg3_ocall(size_t* retval, sgx_dh_msg2_t* dh_msg2, sgx_dh_msg3_t* dh_msg3, int64_t session_id);
+sgx_status_t SGX_CDECL send_dh_msg1_recv_dh_msg2_ocall(size_t* retval, sgx_measurement_t *target_enclave, sgx_dh_msg1_t* dh_msg1, sgx_dh_msg2_t* dh_msg2, int64_t session_id);
+sgx_status_t SGX_CDECL send_dh_msg3_ocall(size_t* retval, sgx_dh_msg3_t* dh_msg3, int64_t session_id);
 
-sgx_status_t SGX_CDECL start_session_ocall(size_t* retval, const char *name, sgx_measurement_t *target_enclave, size_t session_id, size_t *is_server);
-sgx_status_t SGX_CDECL end_session_ocall(size_t* retval, size_t session_id);
+sgx_status_t SGX_CDECL start_session_ocall(size_t* retval, const char *name, sgx_measurement_t *target_enclave, int64_t session_id, size_t *is_server);
+sgx_status_t SGX_CDECL end_session_ocall(size_t* retval, int64_t session_id);
 
-sgx_status_t SGX_CDECL send_msg_ocall(size_t* retval, void* buf, size_t len, size_t session_id);
-sgx_status_t SGX_CDECL recv_msg_ocall(size_t* retval, void* buf, size_t len, size_t session_id);
+sgx_status_t SGX_CDECL send_msg_ocall(size_t* retval, void* buf, size_t len, int64_t session_id);
+sgx_status_t SGX_CDECL recv_msg_ocall(size_t* retval, void* buf, size_t len, int64_t session_id);
 
 sgx_status_t SGX_CDECL fs_init_service_ocall(size_t* retval, size_t num_blocks);
 sgx_status_t SGX_CDECL fs_write_block_ocall(size_t* retval, size_t addr, void* buf, size_t len);
